+++ src/index.d.ts
@@ -1,2 +1,2 @@
-export * from './input_value'
-export * from './radio_value'
+export * from './input__value'
+export * from './radio__value'