+++ src/input__value/index.d.ts
@@ -1,0 +1,6 @@
+export declare function input__value(
+	selector:string|object, parent?:ParentNode|null
+):string|undefined
+export {
+	input__value as input_value,
+}