+++ src/input__value/index.js
@@ -4,7 +4,10 @@ import { _dom2 } from '@ctx-core/dom'
  * @param parent{ParentNode|null}
  * @returns {string|undefined}
  */
-export function input_value(selector, parent) {
+export function input__value(selector, parent) {
 	const dom = _dom2(selector, parent)
 	return dom?.value
+}
+export {
+	input__value as input_value,
 }