+++ src/input_value/index.d.ts
@@ -1,3 +1,0 @@
-export declare function input_value(
-	selector:string|object, parent?:ParentNode|null
-):string|undefined