+++ src/radio__value/index.d.ts
@@ -1,0 +1,4 @@
+export declare function radio__value(name:string):string|undefined
+export {
+	radio__value as radio_value,
+}