+++ src/radio__value/index.js
@@ -2,7 +2,7 @@
  * @param name{string}
  * @returns {string|undefined}
  */
-export function radio_value(name) {
+export function radio__value(name) {
 	const element_a = document.getElementsByName(name)
 	for (let i = 0, l = element_a.length; i < l; i++) {
 		const element = element_a[i]
@@ -11,4 +11,7 @@ export function radio_value(name) {
 		}
 	}
 	return
+}
+export {
+	radio__value as radio_value,
 }